Refactor normalization implementations to infrastructure layer
