+++ docs/architecture/diagrams/component/02-domain-components.mmd
@@ -10,7 +10,7 @@ classDef group              fill:#e6ecef,stroke:#333,stroke-width:1.5px,color:#1
 
 subgraph Domain["Domain Layer"]
     Validation["ValidationService\nbioetl.domain.validation.service"]:::componentPrimary
-    Normalization["NormalizationServiceImpl\nbioetl.domain.transform.impl.normalize"]:::componentPrimary
+    Normalization["NormalizationServiceImpl\nbioetl.infrastructure.transform.impl.normalize"]:::componentPrimary
     Hashing["HashService\nbioetl.domain.transform.hash_service"]:::componentPrimary
     Transformers["TransformerChain\nHash/Index/DbVersion/Fulldate"]:::componentSecondary
     Hasher["HasherImpl\ncanonical serializer"]:::componentSecondary
