+++ src/bioetl/application/pipelines/chembl/base.py
@@ -14,7 +14,7 @@ from bioetl.domain.pipelines.contracts import ErrorPolicyABC, PipelineHookABC
 from bioetl.domain.record_source import RecordSource
 from bioetl.domain.schemas.pipeline_contracts import get_pipeline_contract
 from bioetl.domain.transform.contracts import HashServiceABC, NormalizationServiceABC
-from bioetl.domain.transform.factories import default_normalization_service
+from bioetl.infrastructure.transform.factories import default_normalization_service
 from bioetl.domain.transform.transformers import TransformerABC
 from bioetl.domain.validation.service import ValidationService
