+++ src/bioetl/infrastructure/clients/chembl/provider.py
@@ -10,7 +10,7 @@ from bioetl.domain.transform.contracts import (
     NormalizationConfigProvider,
     NormalizationServiceABC,
 )
-from bioetl.domain.transform.factories import default_normalization_service
+from bioetl.infrastructure.transform.factories import default_normalization_service
 from bioetl.infrastructure.chembl_client import (
     create_client,
     create_extraction_service,
