+++ src/bioetl/infrastructure/transform/__init__.py
@@ -1,5 +1,13 @@
 """Transform infrastructure layer."""
 
-from bioetl.infrastructure.transform.factories import default_normalization_service
+from bioetl.infrastructure.transform.factories import (
+    default_hash_service,
+    default_hasher,
+    default_normalization_service,
+)
 
-__all__ = ["default_normalization_service"]
+__all__ = [
+    "default_hasher",
+    "default_hash_service",
+    "default_normalization_service",
+]
