+++ src/bioetl/infrastructure/transform/impl/__init__.py
@@ -1,7 +1,8 @@
 """Concrete transform implementations."""
 
+from bioetl.infrastructure.transform.impl.hasher import HasherImpl
 from bioetl.infrastructure.transform.impl.normalization_service_impl import (
     NormalizationServiceImpl,
 )
 
-__all__ = ["NormalizationServiceImpl"]
+__all__ = ["HasherImpl", "NormalizationServiceImpl"]
