+++ src/bioetl/infrastructure/transform/impl/normalization_service_impl.py
@@ -8,11 +8,11 @@ from bioetl.domain.transform.contracts import (
     NormalizationConfigProvider,
     NormalizationServiceABC,
 )
-from bioetl.domain.transform.impl.base_normalizer import BaseNormalizationService
-from bioetl.domain.transform.impl.normalize import normalize_scalar
-from bioetl.domain.transform.impl.serializer import serialize_dict, serialize_list
 from bioetl.domain.transform.normalizers import normalize_array, normalize_record
 from bioetl.domain.transform.normalizers.registry import get_normalizer
+from bioetl.infrastructure.transform.impl.base_normalizer import BaseNormalizationService
+from bioetl.infrastructure.transform.impl.normalize import normalize_scalar
+from bioetl.infrastructure.transform.impl.serializer import serialize_dict, serialize_list
 
 
 class NormalizationServiceImpl(NormalizationServiceABC, BaseNormalizationService):
