+++ src/bioetl/infrastructure/transform/impl/normalize.py
@@ -10,8 +10,6 @@ from bioetl.domain.transform.contracts import (
     NormalizationConfigProvider,
     NormalizationServiceABC,
 )
-from bioetl.domain.transform.impl.base_normalizer import BaseNormalizationService
-from bioetl.domain.transform.impl.serializer import serialize_dict, serialize_list
 from bioetl.domain.transform.normalizers import (
     normalize_array,
     normalize_pcid,
@@ -20,6 +18,8 @@ from bioetl.domain.transform.normalizers import (
     normalize_uniprot,
 )
 from bioetl.domain.transform.normalizers.registry import get_normalizer
+from bioetl.infrastructure.transform.impl.base_normalizer import BaseNormalizationService
+from bioetl.infrastructure.transform.impl.serializer import serialize_dict, serialize_list
 
 # Aliases for backward compatibility or convenience
 normalize_pubmed_id = normalize_pmid
