+++ tests/bioetl/domain/transform/test_normalization_prop.py
@@ -9,8 +9,8 @@ import pandas as pd
 from hypothesis import given, settings
 from hypothesis import strategies as st
 
-from bioetl.domain.transform.impl.normalize import normalize_scalar
-from bioetl.domain.transform.impl.serializer import serialize_dict, serialize_list
+from bioetl.infrastructure.transform.impl.normalize import normalize_scalar
+from bioetl.infrastructure.transform.impl.serializer import serialize_dict, serialize_list
 
 
 @settings(suppress_health_check=[], database=None)
